Added the isValidUrl functions into the validations.js file
