+++ web/utils/validation.js
@@ -1,0 +1,10 @@
+const isValidUrl = (string) => {
+  try {
+    new URL(string);
+    return true;
+  } catch (_) {
+    return false;
+  }
+};
+
+export default isValidUrl;
